+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 ### 0.10.0-alpha
 * Migrate to MBrace.Core 0.10 APIs
 * Fix MBraceAzure.ClearAllCloudTasks bug.
+* Client and Service API changed to match MBrace.Core.
+* Added extra options when spawning local workers.
 
 ### 0.7.0-alpha
 * Migrate to MBrace.Runtime.Core 0.9.14 APIs